+++ hc/lib/tests/test_date.py
@@ -6,7 +6,13 @@ from datetime import timezone
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
-from hc.lib.date import format_hms, month_boundaries, seconds_in_month, week_boundaries
+from hc.lib.date import (
+    format_approx_duration,
+    format_hms,
+    month_boundaries,
+    seconds_in_month,
+    week_boundaries,
+)
 
 CURRENT_TIME = datetime(2020, 1, 15, tzinfo=timezone.utc)
 MOCK_NOW = Mock(return_value=CURRENT_TIME)
@@ -36,6 +42,24 @@ class DateFormattingTestCase(TestCase):
 
         s = format_hms(td(seconds=60 * 60))
         self.assertEqual(s, "1 h 0 min 0 sec")
+
+
+class ApproxFormattingTestCase(TestCase):
+    def test_days_work(self) -> None:
+        s = format_approx_duration(td(days=3, hours=6, minutes=12, seconds=24))
+        self.assertEqual(s, "3 days 6 h")
+
+    def test_one_day_works(self) -> None:
+        s = format_approx_duration(td(days=1, hours=6, minutes=12, seconds=24))
+        self.assertEqual(s, "1 day 6 h")
+
+    def test_hours_work(self) -> None:
+        s = format_approx_duration(td(hours=6, minutes=12, seconds=24))
+        self.assertEqual(s, "6 h 12 min")
+
+    def test_minutes_work(self) -> None:
+        s = format_approx_duration(td(minutes=12, seconds=24))
+        self.assertEqual(s, "12 min 24 sec")
 
 
 @patch("hc.lib.date.now", MOCK_NOW)