+++ CHANGELOG.md
@@ -16,6 +16,7 @@ version is 3.10.
 - Improve ntfy notifications (include tags, period, last ping type etc.)
 - Add an "Account closed." confirmation message after closing an account
 - Add monthly uptime percentage display in Check Details page (#773)
+- Increase the precision of calculated downtime duration in check's details and reports
 
 ### Bug Fixes
 - Fix "senddeletionnotices" to recognize "Supporter" subscriptions