+++ CHANGELOG.md
@@ -17,6 +17,7 @@ version is 3.10.
 - Add an "Account closed." confirmation message after closing an account
 - Add monthly uptime percentage display in Check Details page (#773)
 - Increase the precision of calculated downtime duration in check's details and reports
+- Increase bottom margin for modal windows to work around Mobile Safari issue (#899)
 
 ### Bug Fixes
 - Fix "senddeletionnotices" to recognize "Supporter" subscriptions