+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v3.0-dev - Unreleased
+## v3.0 - 2023-10-16
 
 This release drops support of Python 3.9 and below. The minimum required Python
 version is 3.10.